Add draggable attribute to draggable div's in osd tab, and fix data-type mismatch when setting / retrieving data.
Most mobile shims for drag-and-drop assumes that you're either dragging an img-tag or an element with the draggable attribute. When not specified, the event-target will be the nested img instead of the div.

On mobile platforms (atleast iOS), no data is returned when dropping an element if getData() and setData() uses different types.