Second attempt at a fix, observing the change made in #589 is kept.
See commit 6db0b0804833d54329fab610464e8e12f75249df

In testing it was found the response buffer size was 2055 bytes but the
dataSize variable was set to 4096.  The remaining bytes was thus 2048.

An exception is thrown by DataView if the buffer passed to the DataView
constructor is smaller than the length constructor argument.  However,
the exception that is thrown is 'RangeError: Invalid DataView length
undefined' which is misleading.