check if filters are enabled before writing value
I noticed that after every save in the pids tab the values dterm_lowpass_hz and gyro_lowpass_hz are set to 0. This happens because the field is set to 0 when the filter is disabled and is written back with the next save.

Now the value is not changed if the field is disabled and this prevents changing the value from the default value to 0.