refactor CSS references to Open Sans font
The old way of using Open Sans was taking some browser quirk into consideration, instead of asking for the web font such as 'Open Sans' in bold, css would ask for a font called 'open_sansbold'

I also found two instances of 'regularbutton' class being defined for tabs, while the class is styled globally in main.css

This fixes a bug I caused where on Windows you can just ask for a bold version of a font, and it'll do it. macOS doesn't really do this properly, so the langauge selector was hard to see.