Simplify flashingMessage function i18next uses `:` to denote a namespace for messages.
The firmware flashing tab has a `flashingMessage` function that is used to update the bar at the bottom, it can either be passed a message key, or a properly formatted string.

`Loaded Local Firmware: (473178 bytes)` was getting looked up as a key, which generates the following warning

i18next.min.js:1 i18next::translator: key " (473178 bytes)" for namespace "Loaded Local Firmware" for languages "en" won't get resolved as namespace was not yet loaded This means something IS WRONG in your application setup. You access the t function before i18next.init / i18next.loadNamespace / i18next.changeLanguage was done. Wait for the callback or Promise to resolve before accessing it!!!

Proposed fix is to just look at the message for a `:` and excluding those before asking the localization system if its a key.

Revised to just remove the lookup