Work around false error when claiming interface to fix firmware flasing on MacOS The error check after attempting to claim the USB interface was falsely indicating failure on MacOS even though it was working properly. Tried all available versions of NW.js (including beta) and none resolve the problem. The problem only seems to affect MacOS.
The error check after claiming the interface is suppressed for MacOS to work around the problem. Risk is low because even if the climing of the interface failed the worst that could happen is flash fails anyway.

Hopefully there will be some future version of NW.js that resolves the problem and we can remove this workaround later (tested through 0.45.0-beta1 so far).