Workaround for custom defaults on EXST targets by adding known addresses.
Note that this is not a long-term solution as the location of the pointers cannot be verified and if other targets are released that include these memory address and they don't contain valid pointers then this code will insert the custom defaults into the wrong place.

As a side note, the location of the current pointers cannot be *verified* either, as it is only controlled through linker scripts in the firmware.

For a potential long-term solution see the PR for this commit.