Fix Three.js no webgl usage
The version of Three.js used still has support for non WebGL 3D
rendering, but the implementation was broken due to the library
internal changes.

This PR bring back the support adding the needed files and making
changes to make it work. Also in order to improve perfomance for
low end devices, the one missing WebGL only, there is the option
to change detail tolerance on legacy mode.

This mode can now be enable via options flag on main menu.