Update Vue (#4258)
* Update Vue

* Fix sonar

* Fix as many Vue3 issues as possible

The biggest problem is with arrays. Vue3 is not detecting well
reactivity in some arrays because we have the view and the logic in
different files, so we can't access the reactivity proxy in a clean way.

I've added reactivity in the definition of the modules affected as a
workaround, but I'm sure there must be a better way.

---------

Co-authored-by: Miguel Angel Mulero Martinez <migmul@gmail.com>